commontorizon: Add telemetry recipe
This is a community driven project, maintainers are volunteers.
But for now we do not have much "horse power" to maintain this project.
We need to know where put our efforts, so, we need to know
how the project is being used. This recipe add a service that
will send, ANONIMOUSLY DATA, to a server that will collect:

- Board model
- Board arch
- OS release
- Region
- dmesg (only the emerg, alert, crit, err, warning levels)

This is enabled by default, and only in the Common Torizon development
image. Also is opt-out, so, if you do not want to send the data, just
remove the recipe from the image or disable the service.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>