grub-ota-fallback: create recipe to add fallback logic to Grub after OTA
This recipe uses the OSTree change that enables to add custom logic to
grub.cfg. It adds the '99_fallback_logic' to '/etc/ostree.d', so after
an update, if it goes wrong, grub can perform a fallback.

Adds the 'grubenv-create' service, a oneshot service that the idea is to
only run in the first boot ever, since it check the BOOT partition for
the grubenv environment file, and if it's not present, it will create
one with the variables used during fallback with their default values.

Adds an entry into fstab, so that the BOOT partition is mounted and
ready for both Greenboot and grubenv-create.service to operate on Grub
environment file.

And also adds 2 aliases for 'grubenv-edit' tool, for 'fw_printenv' and
'fw_setenv', so that both Aktualizr and Greenboot become compatible with
Grub.

Related-to: TOR-3299

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>