rac: fix some compilation problems
A following compilation error was observed:
|   duplicate key `olpc-cjson` in table `patch.https://github.com/toradex/tough`

that is due to a sed syntax trying to append a 'olpc-cjson' line to
${CARGO_HOME}/config, but it has to delete it before hand, otherwise,
there is a risk when do_compile runs, it will append multiple
'olpc-cjson' lines to ${CARGO_HOME}/config.

also a following compilation error was observed:
|   Unable to update https://github.com/toradex/tough?branch=rac#28c2deb2

that is due to OE's cargo running in offline mode, it can not fetch any
sources through internet, so we should not use 'rac' branch as
SRCREV_tough, it has to match the exact commit hash of tough dependency
in torizon-rac's Cargo.lock file.

Also let's move the code modifying ${CARGO_HOME}/config to
do_configure, as a postfuncs, which makes more sense.

Fix: https://github.com/torizon/meta-toradex-torizon/issues/28

Signed-off-by: Ming Liu <ming.liu@toradex.com>