ostree: make OTA updates possible when composefs is employed
The present commit fixes two issues that prevented OTA updates on the
secure-boot image having composefs (i.e. rootfs protection) enabled.

The first issue is the sysroot being mounted read-only which happens to
be the case now in the secure-boot image (simply because that's the
default assumed by ostree when composefs is enabled). The fix here
explicitly sets the configuration to ask for a rw /sysroot.

The second issue was actually a bug in libostree that caused the
deployment function used by Aktualizr to fail. For fixing this, a patch
to ostree was made. The patch was already accepted upstream so upon the
next ostree version bump it should be possible to drop it from the
layer.

Resolves #40

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>