xenomai: Add distro and recipe for Xenomai 4 on x86
Add a distro named "torizon-xenomai" that select the Xenomai kernel and
set the corresponding kernel command-line args.

Add a recipe that build the "linux-evl" Xenomai 4 kernel fork, which
include the EVL co-processor implementation.

Use the active SLTS kernel 5.10.y branch from Xenomai, because it's the
same version as "linux-intel-lts" for kirkstone in meta-common-torizon.

Set minimal kernel config for Xenomai to work. Further optimization may
be possible, but is not covered.

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>