initramfs-framework: plymouth: do not redirect console
No longer redirect the console with plymouth during boot. With Torizon
OS (whether building with secure-boot support or not), the kernel
command line already has the "quiet" argument in it which prevents most
messages from showing up so redirecting the console seems unnecessary
and it prevents us from showing important information immediately in the
console when needed.

The reason for removing the redirection now is to allow us to provide
feedback on the fsverity enabling process which is a one-time process
that takes a few minutes during the first boot after installation with
Toradex Easy Installer. The feedback part will be added in the following
commit.

Related-to: TOR-3379

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>